chore: move useDefaultConvo
