+++ client/src/hooks/Conversations/index.ts
@@ -1,5 +1,6 @@
 export { default as usePresets } from './usePresets';
 export { default as useGetSender } from './useGetSender';
+export { default as useDefaultConvo } from './useDefaultConvo';
 export { default as useConversation } from './useConversation';
 export { default as useConversations } from './useConversations';
 export { default as useDebouncedInput } from './useDebouncedInput';
